requirements.txt
sqlalchemy~=2.0.25
pip~=24.0
attrs~=23.2.0
distro~=1.7.0
wheel~=0.37.1
cryptography~=3.4.8
Pillow~=9.0.1
Jinja2~=3.1.2
filelock~=3.12.3
redis~=5.0.1
setuptools~=68.2.0
packaging~=21.3
zipp~=1.0.0
future~=0.18.2
platformdirs~=3.10.0
PyJWT~=2.3.0